requirements.txt
python-dotenv
flask
gradio
pyyaml
pdfplumber
reportlab
langchain
langchainhub
langchain-openai
